# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Zoryqon API
  version: 1.0.0
  description: Public market data, authenticated customer trading, provider callbacks, and protected operations.
servers:
  - url: https://api.zoryqon.com
    description: Production
security: []
...
paths:
...
  /v1/markets:
    get:
      tags: [Public]
      operationId: listMarkets
      parameters:
        - { name: query, in: query, schema: { type: string, maxLength: 200 } }
        - { name: category, in: query, schema: { type: string } }
        - {
            name: sort,
            in: query,
            schema: { type: string, enum: [trending, volume, liquidity, newest, ending_soon] },
          }
        - { name: cursor, in: query, schema: { type: string } }
        - {
            name: limit,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 100, default: 24 },
          }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /v1/markets/{marketRef}:
    parameters: [{ $ref: '#/components/parameters/MarketRef' }]
    get:
      tags: [Public]
      operationId: getMarket
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/Error' }
  /v1/markets/{marketRef}/orderbook:
    parameters: [{ $ref: '#/components/parameters/MarketRef' }]
    get:
      tags: [Public]
      operationId: getOrderBook
      parameters:
        - { name: outcomeRef, in: query, schema: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /v1/markets/{marketRef}/trades:
    parameters: [{ $ref: '#/components/parameters/MarketRef' }]
    get:
      tags: [Public]
      operationId: listMarketTrades
      parameters:
        - {
            name: limit,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 500, default: 100 },
          }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /v1/markets/{marketRef}/history:
    parameters: [{ $ref: '#/components/parameters/MarketRef' }]
    get:
      tags: [Public]
      operationId: getMarketHistory
      parameters:
        - { name: outcomeRef, in: query, required: true, schema: { type: string } }
        - {
            name: range,
            in: query,
            schema: { type: string, enum: [1H, 6H, 1D, 1W, 1M, ALL], default: 1D },
          }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /v1/auth/login:
    get:
      tags: [Identity]
      operationId: beginLogin
      parameters:
        - { name: returnTo, in: query, schema: { type: string, default: / } }
      responses:
        '302': { description: Redirect to the configured OIDC provider }
  /v1/auth/callback:
    get:
      tags: [Identity]
      operationId: completeLogin
      parameters:
        - { name: code, in: query, required: true, schema: { type: string } }
        - { name: state, in: query, required: true, schema: { type: string } }
      responses:
        '302':
          { description: Secure cookies set and browser redirected to the validated return path }
  /v1/auth/refresh:
    post:
      tags: [Identity]
      operationId: refreshSession
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Error' }
  /v1/auth/logout:
    post:
      tags: [Identity]
      operationId: logout
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /v1/me:
    get:
      tags: [Customer]
      operationId: getMe
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Error' }
  /v1/me/permissions:
    get:
      tags: [Customer]
      operationId: getMyPermissions
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /v1/balances:
    get:
      tags: [Customer]
      operationId: listBalances
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /v1/positions:
    get:
      tags: [Customer]
      operationId: listPositions
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /v1/trades:
    get:
      tags: [Customer]
      operationId: listMyTrades
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /v1/ledger:
    get:
      tags: [Customer]
      operationId: listMyLedgerTransactions
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /v1/orders:
    get:
      tags: [Customer]
      operationId: listOrders
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
    post:
      tags: [Customer]
      operationId: placeOrder
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlaceOrder' }
      responses:
        '201': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Error' }
  /v1/orders/quote:
    post:
      tags: [Customer]
      operationId: quoteOrder
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrderQuoteRequest' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /v1/orders/{orderRef}:
    delete:
      tags: [Customer]
      operationId: cancelOrder
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters:
        - { name: orderRef, in: path, required: true, schema: { type: string } }
        - { $ref: '#/components/parameters/CsrfToken' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /v1/deposits:
    get:
      tags: [Customer]
      operationId: listDeposits
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
    post:
      tags: [Customer]
      operationId: createDeposit
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDeposit' }
      responses:
        '201': { $ref: '#/components/responses/Success' }
  /v1/withdrawals:
    get:
      tags: [Customer]
      operationId: listWithdrawals
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
    post:
      tags: [Customer]
      operationId: createWithdrawal
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWithdrawal' }
      responses:
        '201': { $ref: '#/components/responses/Success' }
  /v1/withdrawals/{withdrawalRef}/confirm:
    post:
      tags: [Customer]
      operationId: confirmWithdrawal
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters:
        - { name: withdrawalRef, in: path, required: true, schema: { type: string } }
        - { $ref: '#/components/parameters/CsrfToken' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IdempotencyConfirmation' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/verification:
    get:
      tags: [Customer]
      operationId: getVerification
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /v1/verification/start:
    post:
      tags: [Customer]
      operationId: startVerification
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      responses:
        '201': { $ref: '#/components/responses/Success' }
  /v1/sessions:
    get:
      tags: [Customer]
      operationId: listSessions
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /v1/sessions/{sessionRef}:
    delete:
      tags: [Customer]
      operationId: revokeSession
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters:
        - { name: sessionRef, in: path, required: true, schema: { type: string } }
        - { $ref: '#/components/parameters/CsrfToken' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /v1/notification-preferences:
    get:
      tags: [Customer]
      operationId: getNotificationPreferences
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
    put:
      tags: [Customer]
      operationId: updateNotificationPreferences
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /v1/provider-webhooks/payment:
    post:
      tags: [Providers]
      operationId: receivePaymentEvent
      parameters:
        - { name: x-provider-signature, in: header, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '200': { description: Duplicate event acknowledged }
        '202': { description: New verified event processed }
        '400': { $ref: '#/components/responses/Error' }
  /v1/admin/overview:
    get:
      tags: [Operations]
      operationId: getOperationsOverview
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Error' }
...
  /v1/admin/markets:
    get:
      tags: [Operations]
      operationId: listOperationsMarkets
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
    post:
      tags: [Operations]
      operationId: createMarket
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      responses:
        '201': { $ref: '#/components/responses/Success' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC access token verified by issuer, audience, signature, and expiry.
    cookieAuth:
      type: apiKey
      in: cookie
      name: zoryqon_access
  parameters:
    MarketRef:
      name: marketRef
      in: path
      required: true
      schema: { type: string }
    CsrfToken:
      name: x-csrf-token
      in: header
      required: false
      description: Required for cookie-authenticated state changes; bearer-token requests do not use it.
      schema: { type: string }
  schemas:
...
    PlaceOrder:
      allOf:
        - { $ref: '#/components/schemas/OrderQuoteRequest' }
        - type: object
          additionalProperties: false
          required: [type, quoteRef, idempotencyKey]
          properties:
            type: { const: limit }
            quoteRef: { type: string }
            idempotencyKey: { type: string, minLength: 8, maxLength: 128 }
...
    IdempotencyConfirmation:
      type: object
      additionalProperties: false
      required: [idempotencyKey]
      properties:
        idempotencyKey: { type: string, minLength: 8, maxLength: 128 }
  responses:
    Success:
      description: Successful response
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    Error:
      description: Request rejected
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }